refactor: move main files to sub modules
