+++ src/commands/index.js
@@ -1,0 +1,10 @@
+module.exports = {
+  initCommand: require("./init"),
+  addCommand: require("./add"),
+  listCommand: require("./list"),
+  removeCommand: require("./remove"),
+  serveCommand: require("./serve"),
+  homeCommand: require("./home"),
+  validateCommand: require("./validate"),
+  bundleCommand: require("./bundle"),
+};
